Samples/WindowsML/python/SqueezeNetPython/requirements.txt
pillow
numpy
onnxruntime-winml
wasdk-Microsoft.Windows.AI.MachineLearning[all]
wasdk-Microsoft.Windows.ApplicationModel.DynamicDependency.Bootstrap
